WIP switch to postgres. Work on upgrade/migrate scripts.

+++ deployment/native/upgrade_020_030.sh
@@ -4,8 +4,18 @@
 # This script builds a release version and updates files in the install directory as needed.
 # Default install directory is /opt/tms_server. May be overridden using env variable TMS_INSTALL_DIR.
 # User used to build and install TMS may be given on the command line. Default user is "tms"
-
-# TODO Support a test mode where we do not have to run as root and tms_install_user is taken to be current user.
+#
+# Assumptions:
+#  - We are running from a checkout of tms_server github repo.
+#  - Following are installed: rust tool chain (cargo, rustc), SQLite and postgres psql.
+#  - Following env variables are set:
+#    - TMS_DB_HOST     e.g. localhost
+#    - TMS_DB_PORT     e.g. 5431
+#    - TMS_DB_USER     e.g. tms
+#    - TMS_DB_PASSWORD
+#    - POSTGRES_PASSWORD
+#
+# A test mode is supported allowing for execution as a non-root user and tms_install_user is taken to be current user.
 
 PrgName=$(basename "$0")
 USAGE="Usage: $PrgName [ <tms_install_user> | --test ]"
@@ -16,17 +26,24 @@ PRG_RELPATH=$(dirname "$0")
 cd "$PRG_RELPATH"/. || exit
 PRG_PATH=$(pwd)
 
+# Upgrade script is located under deployment/native. Some operations are relative to the top level source directory.
+SRC_DIR=$PRG_PATH/../..
+
 # Check number of arguments
 if [ $# -gt 1 ]; then
   echo "$USAGE"
   exit 1
 fi
+
 TEST_MODE=false
 if [ "$1" == "--test" ]; then
-  echo "Running in test mode"
+  echo "*******************************"
+  echo "     Running in test mode"
+  echo "*******************************"
   TEST_MODE=true
 fi
-# Make sure we are running as root
+
+# Make sure we are running as root or are in test mode
 if [ "$TEST_MODE" == "false" ] && [ "$EUID" != 0 ]; then
   echo "This program must be run as the root user"
   echo "Exiting ..."
@@ -34,11 +51,12 @@ if [ "$TEST_MODE" == "false" ] && [ "$EUID" != 0 ]; then
 fi
 
 # Determine TMS install user
-INSTALL_USR=tms
 if [ "$TEST_MODE" == "true" ]; then
   INSTALL_USR=$USER
 elif [ -n "$1" ]; then
   INSTALL_USR="$1"
+else
+  INSTALL_USR=tms
 fi
 
 # Make sure the specified user for the TMS install exists
@@ -50,7 +68,7 @@ if [ $RET_CODE -ne 0 ]; then
     exit $RET_CODE
 fi
 
-# Set installation directory
+# Set installation directory.
 INSTALL_DEF_DIR="/opt/tms_server"
 if [ -n "$TMS_INSTALL_DIR" ]; then
   INSTALL_DIR="$TMS_INSTALL_DIR"
@@ -77,10 +95,14 @@ fi
 VERS_OLD=$(cat $VERS_FILE)
 
 # Determine new version
-VERS_NEW=$(su - $INSTALL_USR -c 'cd tms_server; cargo pkgid | cut -d "#" -f2')
+if [ "$TEST_MODE" == "true" ]; then
+  VERS_NEW=$('cd ../..; cargo pkgid | cut -d "#" -f2')
+else
+  VERS_NEW=$(su - $INSTALL_USR -c 'cd tms_server; cargo pkgid | cut -d "#" -f2')
+fi
 
 # Set path to built executable
-EXEC_FILE=$PRG_PATH/target/release/tms_server
+EXEC_FILE=$SRC_DIR/target/release/tms_server
 
 # Construct script to be used by install user to build new executable
 echo
@@ -90,7 +112,7 @@ TMP_FILE=$(mktemp)
 # Construct first part of script
 echo "#!/bin/bash" > $TMP_FILE
 # Place various env variables into script
-echo "PRG_PATH=$PRG_PATH" >> $TMP_FILE
+echo "SRC_DIR=$SRC_DIR" >> $TMP_FILE
 echo "INSTALL_DIR=$INSTALL_DIR" >> $TMP_FILE
 echo "VER_OLD=$VERS_OLD" >> $TMP_FILE
 echo "VER_NEW=$VERS_NEW" >> $TMP_FILE
@@ -101,8 +123,8 @@ echo "Upgrading TMS Server from version $VERS_OLD to version $VERS_NEW"
 echo "Install directory: $INSTALL_DIR"
 
 # Build executable
-echo "Building executable from directory: $PRG_PATH"
-cd $PRG_PATH
+echo "Building executable from directory: $SRC_DIR"
+cd $SRC_DIR
 cargo build --release
 EOB
 
@@ -136,8 +158,29 @@ fi
 echo
 echo "===== Stopping TMS service and copying new executable into place"
 echo "========================================================================================="
-systemctl stop tms_server
+if [ "$TEST_MODE" != "true" ]; then
+  systemctl stop tms_server
+fi
 cp $EXEC_FILE $INSTALL_DIR/tms_server
+
+# Update migrations files.
+BAK_TIMESTAMP=`date  +%Y%m%d%H%M%S`
+mv $TMS_HOME/migrations $TMS_HOME/migrations_bak_$BAK_TIMESTAMP
+cp -pr ./resources/migrations $TMS_HOME/
+chmod 0700 $TMS_HOME/migrations
+
+# Before updating version and starting up new tms_server perform the migration from sqlite to postgres
+echo
+echo "===== Migrating DB from sqlite to postgres"
+echo "========================================================================================="
+
+$SRC_DIR/migrate_to_psql/migrate_from_sqlite.sh
+if [ $RET_CODE -ne 0 ]; then
+  echo
+  echo "*************** Error running migration script"
+  echo "Exiting ..."
+  exit $RET_CODE
+fi
 
 # Update version in install dir
 echo "$VERS_NEW" > $VERS_FILE
@@ -146,9 +189,11 @@ echo "$VERS_NEW" > $VERS_FILE
 echo
 echo "===== Starting TMS service"
 echo "========================================================================================="
-systemctl start tms_server
+if [ "$TEST_MODE" != "true" ]; then
+  systemctl start tms_server
+fi
 
-# Remove the temporary file
-rm -f $TMP_FILE
+# TODO Remove the temporary file
+# rm -f $TMP_FILE
 # Switch back to current working directory of invoking user
 cd "$RUN_DIR"
